backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
msal==1.26.0
requests==2.31.0
openpyxl==3.1.2
python-multipart==0.0.6
pytest==7.4.4
httpx==0.26.0
# Auth dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.3
# PostgreSQL for production
psycopg2-binary==2.9.9
